+++ game/g_fire_director.h
@@ -29,14 +29,14 @@ public:
     //! Returns true if the fire director has a valid firing solution
     bool has_solution() const { return _is_valid; }
     //! Return the current firing solution
-    void get_solution(float& bearing, float& elevation) const;
+    void get_solution(double& bearing, double& elevation) const;
 
 protected:
     handle<ship const> _target;
     gun_design const* _gun;
 
-    float _bearing;
-    float _elevation;
+    double _bearing;
+    double _elevation;
     time_delta _time_of_flight;
     bool _is_valid;
 
@@ -44,8 +44,8 @@ protected:
     // fire director (e.g. every ship) with that gun design.
     struct range
     {
-        float elevation;
-        float range;
+        double elevation;
+        double range;
         time_delta time_of_flight;
     };
 
@@ -53,8 +53,8 @@ protected:
     range _table[max_table];
     std::size_t _table_size;
 
-    static constexpr float min_elevation = math::deg2rad(0.f);
-    static constexpr float max_elevation = math::deg2rad(50.f);
+    static constexpr double min_elevation = math::deg2rad(0.0);
+    static constexpr double max_elevation = math::deg2rad(50.0);
 
 protected:
     void update_solution();
@@ -62,7 +62,7 @@ protected:
     //! Fill range data by linear interpolation of the range table. If the given
     //! range falls outside the bounds of the range table then use the nearest
     //! row of the table and return false.
-    bool interpolate_range(float d, range& r) const;
+    bool interpolate_range(double d, range& r) const;
 
     //! Populate the range table by simulating the projectile at varying elevations.
     void populate_table();
