+++ game/g_subsystem.h
@@ -22,12 +22,12 @@ public:
     virtual object_type const& type() const override { return _type; }
     virtual void think() override;
 
-    void damage(object* inflictor, float amount);
-    float damage() const { return _damage; }
-    void repair(float damage_per_second);
+    void damage(object* inflictor, double amount);
+    double damage() const { return _damage; }
+    void repair(double damage_per_second);
 
 protected:
-    float _damage;
+    double _damage;
 
     time_value _damage_time;
     static constexpr time_delta repair_delay = time_delta::from_seconds(0.5f);
@@ -45,26 +45,26 @@ public:
     virtual object_type const& type() const override { return _type; }
     virtual void think() override;
 
-    void set_rudder_target(float rudder_target) { _rudder_target = rudder_target; }
-    float get_rudder_target() const { return _rudder_target; }
-    float get_rudder_angle() const { return _rudder_angle; }
+    void set_rudder_target(double rudder_target) { _rudder_target = rudder_target; }
+    double get_rudder_target() const { return _rudder_target; }
+    double get_rudder_angle() const { return _rudder_angle; }
 
-    void set_speed_target(float speed_target) { _speed_target = speed_target; }
-    float get_speed_target() const { return _speed_target; }
+    void set_speed_target(double speed_target) { _speed_target = speed_target; }
+    double get_speed_target() const { return _speed_target; }
 
 protected:
     //! Linear drag coefficients along longitudinal and transverse axes. These values
     //! also include density and cross-sectional area terms since they are constant.
-    float _linear_drag_coefficient[2];
+    double _linear_drag_coefficient[2];
     //! Torque per angular velocity squared due to drag
-    float _angular_drag_coefficient;
+    double _angular_drag_coefficient;
     //! Inverse moment of inertia
-    float _inverse_inertia;
+    double _inverse_inertia;
 
-    float _rudder_angle;
-    float _rudder_target;
+    double _rudder_angle;
+    double _rudder_target;
 
-    float _speed_target;
+    double _speed_target;
 };
 
 } // namespace game
